// .old_project /packages/cyoda-ui-lib/src/types/types.ts
export interface IndexConfiguration {
  indexId: string;
  indexName: string;
  entityClass: string;
  rangeField: {
    columnName: string;
    columnPath: string;
    type: string;
    clazzType: string;
    decision: null
  };
  noneRangeFields: [
    {
      columnName: string;
      columnPath: string;
      type: string;
      clazzType: string;
      decision: null
    }
  ];
  createDate: string;
  readyForQuerying: boolean;
  returnsList: boolean;
  persisted: boolean;
  decision: string;
}

export interface CompositeIndexExport {
  '@bean': string;
  data: {
    '@meta': string;
    value: CompositeIndexValueExport[]
  };
}

export interface CompositeIndexValueExport {
  '@bean': string;
  entityClass: string;
  name: string;
  rangeColPath: string;
  columns: string[];
}